Move chai timeout helper to separate file
